backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
asyncpg
python-dotenv
python-jose[cryptography]
bcrypt
pydantic[email]
